+++ gui/geometries/OtGeometry.cpp
@@ -82,9 +82,9 @@ void OtGeometryClass::updateBuffers() {
 	clearBuffers();
 
 	bgfx::VertexLayout layout = OtVertex::getVertexLayout();
-	vertexBuffer = bgfx::createVertexBuffer(bgfx::copy(vertices.data(), sizeof(OtVertex) * vertices.size()), layout);
-	triagleIndexBuffer = bgfx::createIndexBuffer(bgfx::copy(triangles.data(), sizeof(uint32_t) * triangles.size()), BGFX_BUFFER_INDEX32);
-	lineIndexBuffer = bgfx::createIndexBuffer(bgfx::copy(lines.data(), sizeof(uint32_t) * lines.size()), BGFX_BUFFER_INDEX32);
+	vertexBuffer = bgfx::createVertexBuffer(bgfx::copy(vertices.data(), sizeof(OtVertex) * (uint32_t) vertices.size()), layout);
+	triagleIndexBuffer = bgfx::createIndexBuffer(bgfx::copy(triangles.data(), sizeof(uint32_t) * (uint32_t) triangles.size()), BGFX_BUFFER_INDEX32);
+	lineIndexBuffer = bgfx::createIndexBuffer(bgfx::copy(lines.data(), sizeof(uint32_t) * (uint32_t) lines.size()), BGFX_BUFFER_INDEX32);
 	refreshBuffers = false;
 }
